fix(domain, nep5): Fix test helper being hung up on first url
